only server side left

+++ public_html/Project/sql/008_create_table_CreditHistory.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS  `CreditHistory`
 (
     `id`         int auto_increment not null,
     `user_id`    int,
-    `credit_diff`      int,
+    `credit_diff`      float,
     `reason`     varchar(255),
     `created`    timestamp default current_timestamp,
     `modified`   timestamp default current_timestamp on update current_timestamp,
